update cpu and ram to float

+++ ReverseProxy/src/reverseproxy/StateTable.java
@@ -17,13 +17,13 @@ public class StateTable {
     private final SocketAddress ip;
     private long totalSent;
     private long losses;
-    private double cpuUsage;
-    private long freeRam;
+    private float cpuUsage;
+    private float freeRam;
     private int rtt;
     private int largbanda;
 
 
-    public StateTable(SocketAddress ip, double cpuUsage, long freeRam, int rtt, int largbanda) {
+    public StateTable(SocketAddress ip, float cpuUsage, float freeRam, int rtt, int largbanda) {
         this.ip = ip;
         this.totalSent = 0;
         this.losses = 0;
@@ -44,11 +44,11 @@ public class StateTable {
         return losses;
     }
 
-    public synchronized double getCpuUsage() {
+    public synchronized float getCpuUsage() {
         return cpuUsage;
     }
 
-    public synchronized long getFreeRam() {
+    public synchronized float getFreeRam() {
         return freeRam;
     }
     
@@ -64,11 +64,11 @@ public class StateTable {
         this.losses = l;
     }
 
-    public synchronized byte setCpuUsage(double cpu) {
+    public synchronized byte setCpuUsage(float cpu) {
         this.cpuUsage = cpu;
     }
 
-    public synchronized int setFreeRam(long ram) {
+    public synchronized int setFreeRam(float ram) {
         this.freeRam = ram;
     }
